Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -4,9 +4,10 @@ char* welcome() {
     return "Welcome to Code Radar!";
 }
 
-int main() {
-    double a;
-    scanf("%lf",&a);
-    printf("You entered: %.4lf", a);
+int main(){
+    float a;
+    scanf("%.4f",&a );
+    printf("You entered: %.4f", a);
     return 0;
 }
+
